Sync abi_date_now with massa-proto change and minor fix

+++ src/types.rs
@@ -752,7 +752,7 @@ pub trait Interface: Send + Sync + InterfaceClone {
         unimplemented!("seed")
     }
 
-    fn date_now(&self) -> Result<f64> {
+    fn date_now(&self) -> Result<u64> {
         unimplemented!("date_now")
     }
 }
